Added POST to IFTTT and Light examples
Added the ability to include 3 values along with the IFTTT trigger in
both the Python version and a new function in the C version titled
cloud_ifttt_trigger_values(). Added Python example of IFTTT integrations
using the Light sensor to adjust the brightness of a Philips Hue light
through IFTTT. C version to follow.