Finished Door Examples, Changed FXOS drivers
Completed the Example_Door.c/py demos that send an IFTTT alert when a
door is opened. Reversed the values of accelerometer and magnetometer in
the FXOS8700CQ C drivers because they were backwards, not sure if the
values are correct as a result but they behave as expected now. Changing
orientation would drastically alter the magnetometer values and placing
a magnet next to the Sensorian would drastically change the
accelerometer values, instead of the other way around. Added a line to
remove new line characters from messages to be printed to the LCD. Added
a print to CloudTools to prevent IFTTT requests from messing up prints
afterwards in the main program. Moved light tolerance factor to be a
global user defined variable.