Added LED to C and fixed C long word printing
Previously weren't explicit functions in the C SensorsInterface to turn
the Sensorian Orange LED on and off, would have had to implicitly call
LED_on or off. Added orange_led_on and orange_led_off to
SensorsInterface.c and SensorsInterface.h. Found a crash bug when trying
to print a string to the screen that contains a "word" longer than the
screen size. Fixed it so that now these words are cut off at the end and
continued on the next line.