Added delay to setupSensorian, Python Doxygen prep
Added a 2 second sleep to the end of setupSensorian in
SensorsInterface.c to remove the need to sleep manually in a project
file after calling setup to ensure all the sensors are ready. Also
changed Python function commenting for the API .py files to docstrings
to allow for proper doxygen parsing. Added comments to variables in the
Python API .py files as well for doxygen.